+++ src/driver.cpp
@@ -301,7 +301,7 @@ void Command_Execute( Command_Options co, string input_file, string output_file 
 	}
 
 	// set base
-	ap.number_of_classes_possible = (unsigned int) pow( G_BASE, ap.max_sequence_size );
+	ap.number_of_classes_possible = (unsigned long long int) pow( G_BASE, ap.max_sequence_size );
 
 	Sequence( *Next_Sequence )(Read_Number &rn, int digits);
 	unsigned int( *Sequence_Value )(Sequence s);
